requirements.txt
openai
langchain-openai
pydantic
langchain-core
langchain
python-dotenv
flask
streamlit
websockets
streamlit_extras
google-cloud
google-cloud-dialogflow-cx
pyaudio
fastapi
uvicorn
PyPDF2
python-multipart
httpx
aiofiles